+++ src/apyfixedarray.cc
@@ -2166,7 +2166,9 @@ APyFixedArray APyFixedArray::from_numbers(
     }
 
     APyFixedArray result(
-        python_iterable_extract_shape(number_seq, "APyFixedArray.from_float"),
+        python_iterable_extract_shape<false, nb::float_, nb::int_, APyFixed, APyFloat>(
+            number_seq, "APyFixedArray.from_float"
+        ),
         int_bits,
         frac_bits,
         bits